requirements.txt
amqp==5.1.1
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
django-bootstrap4==23.2
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-jazzmin==2.6.0
django-mathfilters==1.0.0
django-widget-tweaks==1.5.0
kombu==5.3.2
prompt-toolkit==3.0.39
python-dateutil==2.8.2
redis==5.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
whitenoise==6.6.0
